+++ modules/core/src/lib/fetch/read-file.js
@@ -1,6 +1,6 @@
 import {isBrowser, resolvePath} from '@loaders.gl/loader-utils';
 import * as node from '../../node/read-file-sync.node';
-import {readFileSyncBrowser} from './read-file.browser';
+import {assert} from '@loaders.gl/loader-utils';
 
 // In a few cases (data URIs, node.js) "files" can be read synchronously
 export function readFileSync(url, options = {}) {
@@ -8,5 +8,6 @@ export function readFileSync(url, options = {}) {
   if (!isBrowser && node.readFileSync) {
     return node.readFileSync(url, options);
   }
-  return readFileSyncBrowser(url, options);
+  assert(false);
+  return null;
 }