+++ modules/core/test/index.js
@@ -19,3 +19,5 @@ import './lib/load.spec';
 // TODO - The worker-utils specs test loading, not just worker farm
 // so we run them after util tests, until loading has been split out
 import './worker-utils';
+
+// Deprecated